Update README.md for command line tool extensibility
Clarify the approach for supporting extensibility in command line tools.
Mention the original attempt with `AssemblyLoadContext` and the issues with
`Spectre.Console.Cli`. Highlight the use of [.NET Core Plugins](https://github.com/natemcmaster/DotNetCorePlugins)
for improved functionality. Add an example command for running the application
with specific arguments configured in `launchSettings.json`.